src/tasks/podcast_scraper/src/requirements.txt
requests==2.28.1
pyquery==1.4.3
python-dotenv==0.19.0
psycopg2==2.9.3
sqlalchemy==1.4.23
lxml==4.6.3
requests==2.28.1